+++ src/Gestiones/GestionarPersonas.java
@@ -12,15 +12,15 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.logging.Logger;
 
 /**
  * Gestionar Personas
- * 
+ *
  * @author Plam
  */
 public class GestionarPersonas {
-    private static final Logger LOG = Logger.getLogger(GestionarPersonas.class.getName());
+
+    private static final org.apache.log4j.Logger logger = org.apache.log4j.Logger.getLogger(GestionarPersonas.class);
 
     private List<Persona> listaPersonas;
     private ConexionBBDD conexion;
@@ -43,6 +43,9 @@ public class GestionarPersonas {
      * @return retorna la persona tipo Persona en caso contrario retorna null
      */
     public Persona getPersonaPorNombre(String nombre) {
+        if (!conexion.isConexionExitosa()) {
+            return null;
+        }
         //refresco la lista para que actualize los objetos por si hubo cambios
         refrescarListaPersonas();
         for (Persona persona : listaPersonas) {
@@ -60,6 +63,9 @@ public class GestionarPersonas {
      * @return retorna una lista tipo Persona
      */
     public List<Persona> getListaPersonas() {
+        if (!conexion.isConexionExitosa()) {
+            return new ArrayList<>();
+        }
         //refresco la lista para que actualize los objetos por si hubo cambios
         refrescarListaPersonas();
         return listaPersonas;
@@ -72,6 +78,9 @@ public class GestionarPersonas {
      * @return retorna una persona tipo Persona , en caso contrario retorna null
      */
     public Persona getPersonaPorID(int ID_persona) {
+        if (!conexion.isConexionExitosa()) {
+            return null;
+        }
         //refresco la lista para que actualize los objetos por si hubo cambios
         refrescarListaPersonas();
         for (Persona persona : listaPersonas) {
@@ -89,6 +98,9 @@ public class GestionarPersonas {
      * @return retorna una persona tipo Persona , en caso contrario retorna null
      */
     public Persona getPersonaPorIDFactura(int ID_factura) {
+        if (!conexion.isConexionExitosa()) {
+            return null;
+        }
         String consulta = "select ID_persona from negocio where ID_factura = " + ID_factura;
         ResultSet resultado = conexion.ejecutarStatementSELECT(consulta, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
         try {
@@ -104,12 +116,16 @@ public class GestionarPersonas {
         }
         return null;
     }
-    
+
     /**
      * Metodo para obtener una lista de personas cuyo tipo es CLIENTE
+     *
      * @return return una lista tipo Persona
      */
     public List<Persona> getListaClientes() {
+        if (!conexion.isConexionExitosa()) {
+            return new ArrayList<>();
+        }
         //refresco la lista para que actualize los objetos por si hubo cambios
         refrescarListaPersonas();
         List<Persona> lista = new ArrayList<>();
@@ -123,9 +139,13 @@ public class GestionarPersonas {
 
     /**
      * Metodo para obtener una lista de personas cuyo tipo es PROVEEDOR
+     *
      * @return return una lista tipo Persona
      */
-    public List<Persona> getListaProveedores(){
+    public List<Persona> getListaProveedores() {
+        if (!conexion.isConexionExitosa()) {
+            return new ArrayList<>();
+        }
         //refresco la lista para que actualize los objetos por si hubo cambios
         refrescarListaPersonas();
         List<Persona> lista = new ArrayList<>();
@@ -138,12 +158,17 @@ public class GestionarPersonas {
     }
 
     /**
-     * Metodo para borrar una persona segun el codigo de persona pasado como parametro
+     * Metodo para borrar una persona segun el codigo de persona pasado como
+     * parametro
+     *
      * @param ID_persona el codigo de la persona que se va a borrar
      * @return retorna las filas actualizadas de la base de datos
      */
     public int borrarPersona(int ID_persona) {
         int filas = 0;
+        if (!conexion.isConexionExitosa()) {
+            return -1;
+        }
         String consulta = "delete from persona where ID_PERSONA = " + ID_persona;
         filas = conexion.ejecutarStatementNOSELECT(consulta, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
         //refresco la lista para que actualize los objetos por si hubo cambios
@@ -152,14 +177,18 @@ public class GestionarPersonas {
     }
 
     /**
-     * Metodo para modificar los datos de una persona. 
-     * 
-     * @param p El parametro persona ya contiene los datos nuevos, 
-     * hay que sustituir estos datos con los datos de la persona con el mismo ID en la base de datos
-     * 
-     * @return retorna las filas actualizadas 
+     * Metodo para modificar los datos de una persona.
+     *
+     * @param p El parametro persona ya contiene los datos nuevos, hay que
+     * sustituir estos datos con los datos de la persona con el mismo ID en la
+     * base de datos
+     *
+     * @return retorna las filas actualizadas
      */
     public int modificarPersona(Persona p) {
+        if (!conexion.isConexionExitosa()) {
+            return -1;
+        }
         String consulta = "update persona set nombre = '" + p.getNombre() + "', apellido = '" + p.getApellido() + "', telefono =" + p.getTelefono() + ", direccion = '" + p.getDireccion() + "', email = '" + p.getEmail() + "', type = '" + p.getTipo() + "' where ID_PERSONA = " + p.getCodPersona();
         int filas = conexion.ejecutarStatementNOSELECT(consulta, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
         //refresco la lista para que actualize los objetos por si hubo cambios
@@ -171,6 +200,9 @@ public class GestionarPersonas {
      * Metodo para refrescar la lista de las personas
      */
     private void refrescarListaPersonas() {
+        if (!conexion.isConexionExitosa()) {
+            return;
+        }
         try {
             String consulta = "select * from persona";
             ResultSet resultado = conexion.ejecutarStatementSELECT(consulta, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
@@ -191,36 +223,28 @@ public class GestionarPersonas {
             ex.printStackTrace();
         }
     }
-    
+
     /**
      * Metodo para agregar una persona en la base de datos.
-     * 
+     *
      * @param nombre nombre de la persona
      * @param apellido apellidos de la persona
      * @param direccion su direccion
      * @param telefono su telefono
      * @param email su email
      * @param tipo que tipo es la persona Cliente o Proveedor
-     * @return  retorna las filas actualizadas
+     * @return retorna las filas actualizadas
      */
     public int addPersona(String nombre, String apellido, String direccion, long telefono, String email, String tipo) {
+        if (!conexion.isConexionExitosa()) {
+            return -1;
+        }
         int filas = 0;
-//        try {
-            int ID_Persona = -1;
-//            String consultaNextID = "SELECT max(ID_Persona) FROM persona";
-//            ResultSet resultado = conexion.ejecutarStatementSELECT(consultaNextID, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
-//            if (resultado.next()) {
-//                ID_Persona = resultado.getInt(1);
-//            }
+        String consulta = "insert into persona (nombre, apellido, direccion, telefono, email, type) values ('" + nombre + "', '" + apellido + "', '" + direccion + "', " + telefono + ", '" + email + "', '" + tipo + "')";
+        filas = conexion.ejecutarStatementNOSELECT(consulta, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
+        refrescarListaPersonas();
+        System.out.println("Personas insertadas: " + filas);
 
-            String consulta = "insert into persona (nombre, apellido, direccion, telefono, email, type) values ('" + nombre + "', '" + apellido + "', '" + direccion + "', " + telefono + ", '" + email + "', '" + tipo + "')";
-
-            filas = conexion.ejecutarStatementNOSELECT(consulta, ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
-            refrescarListaPersonas();
-            System.out.println("Personas insertadas: " + filas);
-//        } catch (SQLException e) {
-//            System.out.println("Error en la consulta ...");
-//        }
         return filas;
     }
 }
